docs: phases 4+5 — dedup consolidation + cosmetics batch (D1-D4, C1-C10)
Dedup: Ctrl-to-save collapsed to one owner (unique save-as/OFN and
click-time-GetKeyState content MOVED into conventions/ctrl-to-save.md,
three copies in dialog-conventions became pointers); helper-binary
resolution + Browse-anchoring moved to conventions/tool-resolution.md;
output-dir cascade to its owner; coverage-map's out-of-order revision
narrative trimmed to a summary. dialog-conventions 956 -> 802 lines,
invocation 541 -> 496, all content preserved at its owner.

Cosmetics: SearchHitInfo/SearchInfo/CodePages field aligned to the
header's LONG iSize (page-wide nSize spelling was drift — verified
against vendor source); cpp nOpType LONG -> DWORD (header) +
recompiled; cpp build.bat anchors to its own folder; INDEX refreshed
(search warnings, register status, dedup pointers); frontmatter
normalized across 25 files; register's machine-specific phrasing
neutralized; stale cross-ref titles fixed. C2 was a false positive
(inline code, not a link); C5 decided-keep (raw-markdown admonitions
read fine for the skill's primary consumer). Repaired 4 NUL bytes a
shell-escaping bug introduced into ctrl-to-save.md (repo now verified
NUL-free). Full gate suite green: 8 checks + 2 scaffold suites +
markdownlint 0 issues.

+++ templates/x-tensions/cpp/my_xtension.cpp
@@ -210,7 +210,7 @@ LONG __stdcall XT_About(HWND hParentWnd, void* lpReserved) {
     return 0;
 }
 
-LONG __stdcall XT_Prepare(HANDLE hVolume, HANDLE hEvidence, LONG nOpType, void* lpReserved) {
+LONG __stdcall XT_Prepare(HANDLE hVolume, HANDLE hEvidence, DWORD nOpType, void* lpReserved) {
     g.hVolume = hVolume;
     g.opType  = nOpType;
     g.processed = g.flagged = 0;
@@ -287,7 +287,7 @@ LONG __stdcall XT_ProcessSearchHit(void* info) {
     return 0;
 }
 
-LONG __stdcall XT_Finalize(HANDLE hVolume, HANDLE hEvidence, LONG nOpType, void* lpReserved) {
+LONG __stdcall XT_Finalize(HANDLE hVolume, HANDLE hEvidence, DWORD nOpType, void* lpReserved) {
     wchar_t buf[128];
     swprintf_s(buf, L"done — processed %llu, flagged %llu",
                (unsigned long long)g.processed, (unsigned long long)g.flagged);
